html and css finished and js lack api part
in html and css part, I have finished the basic layout and style of the website, and changes the input type to divs to make just display the letters.
in js part, I have finished the basic logic of the game, but the api part is not finished yet, so the game is not playable now.